requirements.txt
pandas>=1
mysql-connector-python>=8